\name{allneded}
\alias{allneded}
\title{
loads all packages this package depends on
}
\description{loads all packages needed for running the package NPMPM (SweaveListingUtils, graphics)}
\usage{
allneded()
}

\value{
no return value
}

\author{
Nadine Schoene
}
\examples{
function(){

# LOAD PACKAGE FOR GENERATING OF .PDF OUTPUT
library(SweaveListingUtils)

# LOAD PACKAGE FOR PLOTTING
library(graphics) 

  } # END FUNCTION
}
